content/base/test/test_bug600471.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=600471
-->
<head>
  <title>Test for Bug 600471</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=600471">Mozilla Bug 600471</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 600471 **/

var attr = document.createAttribute("foo");
attr.value = "bar";
var node = document.createElement("div");
node.setAttributeNode(attr);

node.setAttribute("foo", "newvalue");
is(attr.firstChild.nodeValue, "newvalue", "the value should be updated");

</script>
</pre>
</body>
</html>
